Separate pie graph into PieGraph component
