+++ expense-splitter/src/components/group/GroupDetail.jsx
@@ -6,6 +6,7 @@ import Button from "../ui/Button";
 import Card from "../ui/Card";
 import Dialog from "../ui/Dialog";
 import { categories } from "../../utils/dummyData";
+import PieGraph from "../widgets/PieGraph";
 
 function GroupDetail() {
   const [deleteID, setDeleteID] = useState(null);
@@ -14,49 +15,14 @@ function GroupDetail() {
 
   // Create reference to dom elements
   const deleteDialogRef = useRef(null);
-  const chartRef = useRef(null);
+
+  const pieChartData = [20, 50, 100, 90, 20, 40, 70];
 
   const { groupId } = useParams();
   const navigate = useNavigate();
 
   const singleGroup = groupData.find((group) => group.id === groupId);
   console.log(singleGroup);
-
-  useEffect(() => {
-    const dataPie = {
-      labels: categories,
-      //   labels: ["JavaScript", "Python", "Ruby"],
-      datasets: [
-        {
-          label: "Categories",
-          data: [20, 50, 100, 90, 20, 40, 70],
-          backgroundColor: [
-            "#ffd373",
-            "#fd8021",
-            "#e05400",
-            "#0073cc",
-            "#003488",
-            "#001d59",
-            "#001524",
-          ],
-          hoverOffset: 4,
-        },
-      ],
-    };
-
-    const configPie = {
-      type: "pie",
-      data: dataPie,
-      options: {},
-    };
-
-    const pieChart = new Chart(chartRef.current, configPie);
-
-    // Cleanup the chart on component unmount
-    return () => {
-      pieChart.destroy();
-    };
-  });
 
   // Closes or opens the dialog
   const toggleDialog = (ref) => {
@@ -126,9 +92,11 @@ function GroupDetail() {
           </p>
         </div>
 
-        <div className="mx-auto max-w-sm overflow-hidden lg:max-w-lg">
-          <canvas className="p-4" ref={chartRef}></canvas>
-        </div>
+        <PieGraph
+          labels={categories}
+          label={"Categories"}
+          data={pieChartData}
+        />
 
         <div>
           {expenses.length > 0 ? (
